requirements.txt
bs4==0.0.1
pandas==0.25.1
Pillow==6.2.0
streamlit==0.57.3
urllib3==1.25.7